Moved shared Address Data to Abstract Class
