+++ src/main/java/com/graphhopper/converter/core/Converter.java
@@ -14,7 +14,7 @@ public class Converter {
     public static GHEntry convertFromNominatim(NominatimEntry response) {
         GHEntry rsp = new GHEntry(response.getOsmId(), response.getGHOsmType(), response.getLat(), response.getLon(),
                 response.getDisplayName(), response.getAddress().country, response.getAddress().getGHCity(),
-                response.getAddress().state, response.getStreetName(), response.getAddress().houseNumber,
+                response.getAddress().state, response.address.getStreetName(), response.getAddress().houseNumber,
                 response.getAddress().postcode, response.getType());
         return rsp;
     }
@@ -70,7 +70,7 @@ public class Converter {
 
         GHEntry rsp = new GHEntry(osmId, type, response.getGeometry().lat, response.getGeometry().lng,
                 response.getFormatted(), response.getComponents().country, response.getComponents().getGHCity(),
-                response.getComponents().state, response.getStreetName(), response.getComponents().houseNumber,
+                response.getComponents().state, response.components.getStreetName(), response.getComponents().houseNumber,
                 response.getComponents().postcode, response.getComponents().type);
 
         return rsp;
